chore(faceted-search): Move ClearBasicSearch into BasicSearch
