--- packages/faceted-search/src/components/ClearBasicSearch/ClearBasicSearch.scss
@@ -1,10 +1,0 @@
-.tc-faceted-clear-basic {
-	border-right: 0.1rem solid $gallery;
-	margin-right: 1.5rem;
-
-	&-button {
-		padding: 0 0.5rem;
-		in-height: 2.5rem;
-		line-height: 2.5rem;
-	}
-}
